Move tag class to buisness logic package
