+++ src/main/java/com/example/demo/models/businesslogic/PositiveLaw.java
@@ -1,7 +1,6 @@
 package com.example.demo.models.businesslogic;
 
 import com.example.demo.models.entities.LawFormulation;
-import com.example.demo.models.entities.Tag;
 
 import java.util.List;
